Refactored SaveLoadManager extracting serializable state management into SerializableStateManager class
This new class will be extended to contain a scene cache to persist building interiors & player owned areas.
