+++ Assets/Scripts/Game/Serialization/SerializableActionObject.cs
@@ -1,4 +1,4 @@
-﻿// Project:         Daggerfall Tools For Unity
+// Project:         Daggerfall Tools For Unity
 // Copyright:       Copyright (C) 2009-2016 Daggerfall Workshop
 // Web Site:        http://www.dfworkshop.net
 // License:         MIT License (http://www.opensource.org/licenses/mit-license.php)
@@ -37,13 +37,13 @@ namespace DaggerfallWorkshop.Game.Serialization
         void Start()
         {
             if (LoadID != 0)
-                SaveLoadManager.RegisterSerializableGameObject(this);
+                SaveLoadManager.StateManager.RegisterSerializableGameObject(this);
         }
 
         void OnDestroy()
         {
             if (LoadID != 0)
-                SaveLoadManager.DeregisterSerializableGameObject(this);
+                SaveLoadManager.StateManager.DeregisterSerializableGameObject(this);
         }
 
         #endregion
