+++ Assets/Scripts/Game/Serialization/SerializableEnemy.cs
@@ -1,4 +1,4 @@
-﻿// Project:         Daggerfall Tools For Unity
+// Project:         Daggerfall Tools For Unity
 // Copyright:       Copyright (C) 2009-2016 Daggerfall Workshop
 // Web Site:        http://www.dfworkshop.net
 // License:         MIT License (http://www.opensource.org/licenses/mit-license.php)
@@ -51,18 +51,18 @@ namespace DaggerfallWorkshop.Game.Serialization
                 // Only fixing for enemies now - will look for a better solution in the future
                 if (enemy && GameManager.Instance.PlayerEnterExit.IsPlayerInsideDungeon)
                 {
-                    if (SaveLoadManager.Instance.ContainsEnemy(enemy.LoadID))
+                    if (SaveLoadManager.StateManager.ContainsEnemy(enemy.LoadID))
                         enemy.LoadID++;
                 }
 
-                SaveLoadManager.RegisterSerializableGameObject(this);
+                SaveLoadManager.StateManager.RegisterSerializableGameObject(this);
             }
         }
 
         void OnDestroy()
         {
             if (LoadID != 0)
-                SaveLoadManager.DeregisterSerializableGameObject(this);
+                SaveLoadManager.StateManager.DeregisterSerializableGameObject(this);
         }
 
 #if UNITY_EDITOR
