+++ Assets/Scripts/Game/Serialization/SerializableStateManager.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: ea0b4248a3937f04b9b3f6dfb4840586
+timeCreated: 1517062504
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
